Fix bug in solve(...)
Previously the vector `p` was not redistributed to `redistLHS` each time
the `solve(...)` is called when there are more CPU ranks than GPUs.
This caused that the AmgX solver always solved the system with the
solution it got in the last calling since the second call, regardless
what the real LHS vectors users input. For PetIBM, there was no problem
even the bug is not fixed, because, in PetIBM, each calling of the AmgX
solver, we indeed want to solve the system with the solution we got in
the last calling, i.e. the last time step.


Former-commit-id: 750b2d05449caa85fdc089795e30768230488085