Change the way to handle all-Neumann BCs
In matrix A, we previously add 1 to the entry A11 and add the exact
solution of p1 to RHS vector. However, we found this makes the matrix A
indefinite, and hence it sometimes caused some problem when using CG solver.
Now we minus 1 to the entry A11 instead and minus exact solution of p1
from RHS vector.